changes to make map search working for public users
